Updated lambda changes to - Remove the lambda helper and replace it with a generic document uploader - reduce the number of calls to the generic upload_document for simplicity - refactor tests to fit the new generic upload document and add more test coverage to existing code
